gh: add pull request template (#58)

+++ src/rpm/headers/types.rs
@@ -152,7 +152,7 @@ impl From<FileMode> for u32 {
 
 impl From<FileMode> for u16 {
     fn from(mode: FileMode) -> Self {
-        mode.raw_mode() as u16
+        mode.raw_mode()
     }
 }
